Extract pseudopackage for oidc
