+++ authentication.html
@@ -5,7 +5,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1"/>
         <title>ESS Permissions Service</title>
         <link rel="shortcut icon" type="image/png" href="./logo.png"/>
-        <script src="./scripts/authentication.js" type="module" defer></script>
+        <script src="./packages/oidc/authentication.js" type="module" defer></script>
     </head>
     <body>
     </body>
