+++ packages/oidc/authentication.js
@@ -1,4 +1,4 @@
-import {Oidc} from "./Vocabulary.js"
+import {Oidc} from "../../scripts/Vocabulary.js"
 import {OidcClient} from "./OidcClient.js"
 
 addEventListener("load", onLoad)
