+++ packages/oidc/index.js
@@ -1,0 +1,4 @@
+import "../dialog/index.js"
+
+export {OidcCredentialManager} from "./OidcCredentialManager.js"
+export {OidcClient} from "./OidcClient.js"
